auth is done on same process now and cleaned up code
